feat: move ProjectFormSchema to a separate schema file and update imports in project pages
