+++ public/robots.txt
@@ -7,4 +7,4 @@ Host: https://asterixhdev.vercel.app
 
 # Sitemaps
 Sitemap: https://asterixhdev.vercel.app/sitemap.xml
-Sitemap: https://asterixhdev.vercel.app/server-sitemap.xml
+Sitemap: https://asterixhdev.vercel.app/sitemap-0.xml
