+++ public/sitemap-0.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>https://asterixhdev.vercel.app/icon.jpg</loc><lastmod>2025-04-12T19:14:21.479Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://asterixhdev.vercel.app</loc><lastmod>2025-04-12T19:14:21.479Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://asterixhdev.vercel.app/projects</loc><lastmod>2025-04-12T19:14:21.479Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://asterixhdev.vercel.app/admin/files</loc><lastmod>2025-04-12T19:14:21.479Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://asterixhdev.vercel.app/admin</loc><lastmod>2025-04-12T19:14:21.479Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://asterixhdev.vercel.app/admin/projects/new</loc><lastmod>2025-04-12T19:14:21.479Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://asterixhdev.vercel.app/admin/projects</loc><lastmod>2025-04-12T19:14:21.479Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://asterixhdev.vercel.app/icon.jpg</loc><lastmod>2025-04-12T22:06:10.080Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://asterixhdev.vercel.app</loc><lastmod>2025-04-12T22:06:10.080Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://asterixhdev.vercel.app/projects</loc><lastmod>2025-04-12T22:06:10.080Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://asterixhdev.vercel.app/admin/files</loc><lastmod>2025-04-12T22:06:10.080Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://asterixhdev.vercel.app/admin</loc><lastmod>2025-04-12T22:06:10.080Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://asterixhdev.vercel.app/admin/projects/new</loc><lastmod>2025-04-12T22:06:10.080Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://asterixhdev.vercel.app/admin/projects</loc><lastmod>2025-04-12T22:06:10.080Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
 </urlset>
