+++ public/sitemap.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <sitemapindex xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
 <sitemap><loc>https://asterixhdev.vercel.app/sitemap-0.xml</loc></sitemap>
-<sitemap><loc>https://asterixhdev.vercel.app/server-sitemap.xml</loc></sitemap>
+<sitemap><loc>https://asterixhdev.vercel.app/sitemap-0.xml</loc></sitemap>
 </sitemapindex>
